requirements.txt
streamlit
transformers
torch
scikit-learn
google-generativeai
langchain
langchain-google-genai
openai>=1.2.0